feat(landing-page): migrate landing page components to widgets architecture
- Move components from app directory to widgets/landing-page/ui
- Refactor component naming for consistency (kebab-case files)
- Update component exports with named exports
- Improve code formatting and structure
- Fix error handling in Loader component
- Standardize component props interfaces
