--- app/[language]/(guest)/(default-layout)/_home/BubbleDialog/index.ts
@@ -1,1 +1,0 @@
-export { BubbleDialog } from './BubbleDialog';
