--- app/[language]/(guest)/(default-layout)/_home/CallToAction/CTASection.tsx
@@ -1,44 +1,0 @@
-import { cn } from '@/utils/cn';
-import { AuthButton } from '@/contexts/Auth';
-import Image from 'next/image';
-
-interface CTASectionProps {
-  className?: string;
-}
-
-export function CTASection({ className }: CTASectionProps) {
-
-  return (
-    <section 
-      className={cn(
-        'relative flex flex-col justify-center items-center min-h-[366px] my-20 px-6',
-        'overflow-hidden',
-        className
-      )}
-    >
-      <Image 
-        src="/assets/landing-page/bg-island.svg" 
-        alt="" 
-        fill
-        className="object-cover md:object-contain object-center z-0"
-        aria-hidden="true"
-      />
-      <h2 
-        className="relative z-10 text-[20px] md:text-[24px] font-semibold text-primary-darker text-center my-4 leading-tight"
-      >
-        準備好重新打造<br />
-        你喜歡的學習生活了嗎？
-      </h2>
-      <div
-        className="relative z-10"
-      >
-        <AuthButton 
-          variant="ctaOrange"
-          size="huge"
-        >
-          立即加入
-        </AuthButton>
-      </div>
-    </section>
-  );
-}
