--- app/[language]/(guest)/(default-layout)/_home/CallToAction/index.ts
@@ -1,1 +1,0 @@
-export { CTASection } from './CTASection';
