--- app/[language]/(guest)/(default-layout)/_home/Features/index.ts
@@ -1,2 +1,0 @@
-export { FeatureCard } from './FeatureCard';
-export { FeatureGrid } from './FeatureGrid';
