--- app/[language]/(guest)/(default-layout)/_home/Functions/index.ts
@@ -1,2 +1,0 @@
-export { FunctionCarousel } from './FunctionCarousel';
-export { FunctionCard } from './FunctionCard';
