--- app/[language]/(guest)/(default-layout)/_home/KeyVision/index.ts
@@ -1,2 +1,0 @@
-export { KeyVision } from './Keyvision';
-export { default as LottieHero } from './LottieHero';
