--- app/[language]/(guest)/(default-layout)/_home/Loader/index.ts
@@ -1,1 +1,0 @@
-export { Loader } from './Loader';
