--- app/[language]/(guest)/(default-layout)/_home/PersonalityTest/index.ts
@@ -1,1 +1,0 @@
-export { PersonalitySection } from './PersonalitySection';
