--- app/[language]/(guest)/(default-layout)/_home/Plans/index.ts
@@ -1,1 +1,0 @@
-export { PlanSection } from './PlanSection';
