--- app/[language]/(guest)/(default-layout)/_home/Presentation/index.ts
@@ -1,1 +1,0 @@
-export { PresentationSection } from './PresentationSection';
