--- app/[language]/(guest)/(default-layout)/_home/Slogan/index.ts
@@ -1,1 +1,0 @@
-export { SloganSection } from './SloganSection';
