--- app/[language]/(guest)/(default-layout)/_home/Testimonials/index.ts
@@ -1,2 +1,0 @@
-export { TestimonialCard } from './TestimonialCard';
-export { TestimonialMarquee } from './TestimonialMarquee';
