--- app/[language]/(guest)/(default-layout)/_home/Videos/index.ts
@@ -1,2 +1,0 @@
-export { VideoItem } from './VideoItem';
-export { VideoSection } from './VideoSection';
