--- app/[language]/(guest)/(default-layout)/_home/index.ts
@@ -1,12 +1,0 @@
-export { Loader } from './Loader';
-export { KeyVision } from './KeyVision';
-export { SloganSection } from './Slogan';
-export { FeatureCard, FeatureGrid } from './Features';
-export { PresentationSection } from './Presentation';
-export { BubbleDialog } from './BubbleDialog';
-export { VideoItem, VideoSection } from './Videos';
-export { FunctionCard, FunctionCarousel } from './Functions';
-export { PlanSection } from './Plans';
-export { TestimonialCard, TestimonialMarquee } from './Testimonials';
-export { CTASection } from './CallToAction';
-export { PersonalitySection } from './PersonalityTest';
