+++ widgets/landing-page/ui/index.ts
@@ -1,0 +1,12 @@
+export { KeyVision } from './key-vision';
+export { CallToActionSection } from './call-to-action-section';
+export { FeatureGrid } from './feature-grid';
+export { FunctionCarousel } from './function-carousel';
+export { Loader } from './loader';
+export { PersonalitySection } from './personality-section';
+export { PlanSection } from './plan-section';
+export { PresentationSection } from './presentation-section';
+export { SloganSection } from './slogan-section';
+export { TestimonialMarquee } from './testimonial-marquee';
+export { TypewriterBubble } from './typewriter-bubble';
+export { VideoSection } from './video-section';
