+++ widgets/landing-page/ui/key-vision/index.ts
@@ -1,0 +1,1 @@
+export { KeyVision } from './key-vision';
