+++ widgets/landing-page/ui/slogan-section.tsx
@@ -7,13 +7,15 @@ interface SloganSectionProps {
 }
 
 export function SloganSection({ className }: SloganSectionProps) {
-
   return (
-    <section className={cn('slogan-section bg-primary-palest px-6  text-basic-400  relative min-h-[195px] md:min-h-[200px]', className)}>
+    <section
+      className={cn(
+        'slogan-section relative min-h-[195px] bg-primary-palest px-6 text-basic-400 md:min-h-[200px]',
+        className
+      )}
+    >
       {/* 背景島嶼裝飾圖片 */}
-      <div
-        className="absolute top-16 md:-top-4 left-1/2 z-10 transform -translate-x-1/2"
-      >
+      <div className="absolute left-1/2 top-16 z-10 -translate-x-1/2 md:-top-4">
         <Image
           src="/assets/landing-page/deco-island.svg"
           alt="島嶼裝飾"
@@ -22,11 +24,9 @@ export function SloganSection({ className }: SloganSectionProps) {
           data-preload
         />
       </div>
-      
+
       {/* 文字內容 */}
-      <div
-        className="absolute top-32 left-1/2 z-10 w-full transform -translate-x-1/2 -translate-y-1/2"
-      >
+      <div className="absolute left-1/2 top-32 z-10 w-full -translate-x-1/2 -translate-y-1/2">
         <SectionHeader
           title={
             <>
@@ -35,7 +35,6 @@ export function SloganSection({ className }: SloganSectionProps) {
               透過交流與分享，連結成群島
             </>
           }
-
           subtitle="Where personal growth meets collective wisdom!"
           variant="dark"
           alignment="center"
@@ -43,8 +42,6 @@ export function SloganSection({ className }: SloganSectionProps) {
           subtitleClassName="text-basic-400 italic"
         />
       </div>
-
-      
     </section>
   );
 }
